# docs/swagger.yaml
definitions:
...
  dto.AddressDTO:
    properties:
      city_id:
        type: string
      city_name:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      label:
        type: string
      lat:
        type: number
      line1:
        type: string
      line2:
        type: string
      lng:
        type: number
      pincode:
        type: string
      serviceable:
        description: true if the pincode resolved to a serving city
        type: boolean
    type: object
...
  dto.MeDTO:
    properties:
      current_city_id:
        type: string
      current_city_name:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
...
  dto.MoneyDTO:
    properties:
      display:
        type: string
      paise:
        type: integer
    type: object
...
  dto.NotificationPreferencesDTO:
    properties:
      order_updates:
        type: boolean
      price_alerts:
        type: boolean
      promotions:
        type: boolean
    type: object
...
  dto.ProductCardDTO:
    properties:
      brand_name:
        type: string
      default_variant_id:
        type: string
      image_url:
        type: string
      name:
        type: string
      product_id:
        type: string
      starting_price:
        $ref: '#/definitions/dto.MoneyDTO'
    type: object
  dto.ProductCardListDTO:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.ProductCardDTO'
        type: array
      meta:
        $ref: '#/definitions/query.Meta'
    type: object
...
  dto.Response:
    description: Generic response envelope wrapping every API response.
    properties:
      code:
        description: Machine-readable error/status code, present on failures
        type: string
      data:
        description: Response payload, shape depends on the endpoint
      message:
        description: Human-readable message describing the result
        type: string
      status:
        description: Outcome of the request, e.g. "success" or "failed"
        type: string
    type: object
...
  dto.UserListItemDTO:
    description: Per-row shape returned by GET /v1/users.
    properties:
      created_at:
        description: Creation timestamp, RFC3339
        type: string
      phone:
        description: Phone number (login identity)
        type: string
      status:
        description: Status of the client
        enum:
        - active
        - disabled
        - banned
        type: string
      user_id:
        description: Unique identifier for the client
        type: string
      user_name:
        description: Name of the client
        type: string
    type: object
  dto.UserListResponseDTO:
    description: Paginated envelope for GET /v1/users.
    properties:
      items:
        description: Page of users matching the current filters
        items:
          $ref: '#/definitions/dto.UserListItemDTO'
        type: array
      meta:
        allOf:
        - $ref: '#/definitions/query.Meta'
        description: Pagination metadata
    type: object
...
  query.Meta:
    properties:
      page:
        type: integer
      pages:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  v1.BanUserRequest:
    properties:
      user_id:
        type: string
    required:
    - user_id
    type: object
...
  v1.addressRequest:
    properties:
      is_default:
        type: boolean
      label:
        type: string
      lat:
        type: number
      line1:
        type: string
      line2:
        type: string
      lng:
        type: number
      pincode:
        type: string
    required:
    - line1
    - pincode
    type: object
  v1.fcmDeleteRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  v1.fcmTokenRequest:
    properties:
      platform:
        type: string
      token:
        type: string
    required:
    - token
    type: object
...
  v1.updateMeRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
info:
  contact:
    name: API Support
  description: This is the API documentation for the Grocerics backend service.
  title: Grocerics API
  version: "1.1"
paths:
...
  /v1/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MeDTO'
              type: object
      security:
      - BearerAuth: []
      summary: Get my profile
      tags:
      - profile
    patch:
      consumes:
      - application/json
      parameters:
      - description: profile fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.updateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MeDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update my profile
      tags:
      - profile
  /v1/me/addresses:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.AddressDTO'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List my delivery addresses
      tags:
      - profile
    post:
      consumes:
      - application/json
      description: Saves an address; the pincode resolves to a serving city. Setting
        is_default updates the user's current city.
      parameters:
      - description: address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.addressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AddressDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Add a delivery address
      tags:
      - profile
  /v1/me/addresses/{id}:
    delete:
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Delete a delivery address
      tags:
      - profile
    patch:
      consumes:
      - application/json
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: string
      - description: address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.addressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AddressDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update a delivery address
      tags:
      - profile
  /v1/me/fcm-token:
    delete:
      consumes:
      - application/json
      parameters:
      - description: device token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.fcmDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Remove a device push token
      tags:
      - profile
    post:
      consumes:
      - application/json
      parameters:
      - description: device token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.fcmTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Register a device push token
      tags:
      - profile
  /v1/me/notification-preferences:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.NotificationPreferencesDTO'
              type: object
      security:
      - BearerAuth: []
      summary: Get my notification preferences
      tags:
      - profile
    put:
      consumes:
      - application/json
      parameters:
      - description: preferences
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.NotificationPreferencesDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.NotificationPreferencesDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update my notification preferences
      tags:
      - profile
...
  /v1/search:
    get:
      description: Full-text search over the Groceric catalog (min 2 chars). Never
        hits the platform APIs.
      parameters:
      - description: search term (min 2 chars)
        in: query
        name: q
        required: true
        type: string
      - description: page number (default 1)
        in: query
        name: page
        type: integer
      - description: page size, max 100 (default 20)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductCardListDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Search products (curated catalog)
      tags:
      - consumer
...
  /v1/users:
    get:
      description: Paginated list of mobile clients (name, phone, status).
      parameters:
      - description: page number (default 1)
        in: query
        name: page
        type: integer
      - description: page size, max 100 (default 20)
        in: query
        name: page_size
        type: integer
      - description: 'sort column: created_at | name (default created_at)'
        in: query
        name: sort
        type: string
      - description: asc | desc (default desc)
        in: query
        name: order
        type: string
      - description: 'filter by status: active | disabled | banned'
        in: query
        name: status
        type: string
      - description: search by name or phone (max 128 chars)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Users list
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserListResponseDTO'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: List clients
      tags:
      - Users
  /v1/users/ban:
    post:
      consumes:
      - application/json
      description: Sets the client's status to "banned". Irreversible.
      parameters:
      - description: Ban User Request
        in: body
        name: BanUserRequest
        required: true
        schema:
          $ref: '#/definitions/v1.BanUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User banned successfully
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Ban a client
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"